+++ Source/SwiftLintBuiltInRules/Rules/Idiomatic/OneDelarationPerFileRule.swift
@@ -1,78 +1,35 @@
-import Foundation
 import SwiftSyntax
 
 @SwiftSyntaxRule
-struct OneDelarationPerFileRule: Rule {
-    var configuration = SeverityConfiguration<Self>(.error)
+struct OneDelarationPerFileRule: OptInRule {
+    var configuration = SeverityConfiguration<Self>(.warning)
 
     static let description = RuleDescription(
         identifier: "one_declaration_per_file",
         name: "One Declaration Per File",
-        description: "One declaration per file is allowed, extensions are an exception",
+        description: "Only a single declaration is allowed in a file",
         kind: .idiomatic,
         nonTriggeringExamples: [
             Example("""
-                    class Car {
-                        var make: String
-                        var model: String
-                        init(make: String, model: String) {
-                            self.make = make
-                            self.model = model
-                        }
-                    }
+                    class Foo {}
                     """),
             Example("""
-                    class Car {
-                        var make: String
-                        var model: String
-                        init(make: String, model: String) {
-                            self.make = make
-                            self.model = model
-                        }
-                    }
-                    extension Car {
-                        func drive() {
-                        }
-                        func stop() {
-                        }
-                    }
+                    class Foo {}
+                    extension Foo {}
                     """)
         ],
         triggeringExamples: [
             Example("""
-                    class Car {
-                        var make: String
-                        var model: String
-                        init(make: String, model: String) {
-                            self.make = make
-                            self.model = model
-                        }
-                    }
-                    class Bike {
-                        func ride() {
-                        }
-                        func stop() {
-                        }
-                    }
+                    class Foo {}
+                    ↓class Bar {}
                     """),
             Example("""
-                    protocol Identifiable {
-                        var identifier: String { get }
-                    }
-                    enum IdentifiableTypes: String {
-                        case linear, composite
-                    }
+                    protocol Foo {}
+                    ↓enum Bar {}
                     """),
             Example("""
-                    struct BasicProfile {
-                        var id: Int
-                        var name: String
-                    }
-                    struct DetailedProfile {
-                        var basic: BasicProfile
-                        var age: String
-                        var genderRaw: String
-                    }
+                    struct Foo {}
+                    ↓struct Bar {}
                     """)
         ]
     )
@@ -80,37 +37,31 @@ struct OneDelarationPerFileRule: Rule {
 
 private extension OneDelarationPerFileRule {
     final class Visitor: ViolationsSyntaxVisitor<ConfigurationType> {
-        var declarationsCount: Int = 0
-        override var skippableDeclarations: [any DeclSyntaxProtocol.Type] {
-            return .allExcept(ClassDeclSyntax.self, StructDeclSyntax.self, EnumDeclSyntax.self, ProtocolDeclSyntax.self)
+        var declarationVisited = false
+        override var skippableDeclarations: [any DeclSyntaxProtocol.Type] { return .all }
+
+        override func visitPost(_ node: ClassDeclSyntax) {
+            appendViolationIfNeeded(node: node.classKeyword)
+            declarationVisited = true
         }
 
-        override func visit(_ node: ClassDeclSyntax) -> SyntaxVisitorContinueKind {
-            declarationsCount += 1
-            appendViolationIfNeeded(node: node.name)
-            return .skipChildren
+        override func visitPost(_ node: StructDeclSyntax) {
+            appendViolationIfNeeded(node: node.structKeyword)
+            declarationVisited = true
         }
 
-        override func visit(_ node: StructDeclSyntax) -> SyntaxVisitorContinueKind {
-            declarationsCount += 1
-            appendViolationIfNeeded(node: node.name)
-            return .skipChildren
+        override func visitPost(_ node: EnumDeclSyntax) {
+            appendViolationIfNeeded(node: node.enumKeyword)
+            declarationVisited = true
         }
 
-        override func visit(_ node: EnumDeclSyntax) -> SyntaxVisitorContinueKind {
-            declarationsCount += 1
-            appendViolationIfNeeded(node: node.name)
-            return .skipChildren
-        }
-
-        override func visit(_ node: ProtocolDeclSyntax) -> SyntaxVisitorContinueKind {
-            declarationsCount += 1
-            appendViolationIfNeeded(node: node.name)
-            return .skipChildren
+        override func visitPost(_ node: ProtocolDeclSyntax) {
+            appendViolationIfNeeded(node: node.protocolKeyword)
+            declarationVisited = true
         }
 
         func appendViolationIfNeeded(node: TokenSyntax) {
-            if declarationsCount > 1 {
+            if declarationVisited {
                 violations.append(node.positionAfterSkippingLeadingTrivia)
             }
         }